requirements.txt
django
sentence-transformers
faiss-cpu
numpy
pandas
